refactor(plugin-react-refresh): using tsc to build
