+++ packages/rspack-plugin-react-refresh/options.json
