+++ packages/rspack-plugin-react-refresh/package.json
@@ -3,22 +3,25 @@
   "version": "0.3.8",
   "license": "MIT",
   "description": "React refresh plugin for rspack",
-  "main": "src/index.js",
-  "types": "src/index.d.ts",
+  "main": "dist/index.js",
+  "types": "dist/index.d.ts",
   "exports": {
     ".": {
-      "types": "./src/index.d.ts",
-      "default": "./src/index.js"
+      "types": "./dist/index.d.ts",
+      "default": "./dist/index.js"
     },
     "./react-refresh": "./client/reactRefresh.js",
     "./react-refresh-entry": "./client/reactRefreshEntry.js"
   },
   "scripts": {
+    "build": "rimraf dist/ && tsc -b ./tsconfig.build.json --force",
+    "dev": "tsc -b -w",
     "test": "jest --runInBand"
   },
   "files": [
     "client",
-    "src"
+    "dist",
+    "options.json"
   ],
   "publishConfig": {
     "access": "public"
